+++ packages/failwell/vitest.config.ts
@@ -1,0 +1,19 @@
+import tsconfigPaths from 'vite-tsconfig-paths';
+import { defineConfig } from 'vitest/config';
+
+const testFiles = ['./src/**/*.test.{js,ts}'];
+
+export default defineConfig({
+  plugins: [tsconfigPaths()],
+  test: {
+    globals: true,
+    environment: 'node',
+    passWithNoTests: false,
+    include: testFiles,
+    // setupFiles: './config/tests/setupVitest.ts',
+    coverage: {
+      reporter: ['text', 'clover'],
+      extension: ['js', 'jsx', 'ts', 'tsx'],
+    },
+  },
+});